Move the existing tests to the alternator subpackage
