+++ tests/src/test/scala/com/scylladb/migrator/alternator/MigratorSuite.scala
@@ -1,4 +1,4 @@
-package com.scylladb.migrator
+package com.scylladb.migrator.alternator
 
 import com.amazonaws.auth.{AWSCredentials, AWSStaticCredentialsProvider, BasicAWSCredentials}
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
